requirements.txt
pandas
numpy
scikit-learn
xgboost
shap
supabase
fastapi
uvicorn
python-dotenv
imbalanced-learn
matplotlib
seaborn